bullets; moved to float coords instead of int

+++ CivilWar/CivilWar Server/Bullet.h
@@ -10,8 +10,8 @@ public:
 
 
 	Point getCenter() const { return m_center; }
-	int getX() const { return m_center.x; }
-	int getY() const { return m_center.y; }
+	float getX() const { return m_center.x; }
+	float getY() const { return m_center.y; }
 	Radius getRadius() const { return m_radius; }
 	DirectionX getDirectionX() const { return m_dirX; }
 	DirectionY getDirectionY() const { return m_dirY; }
@@ -19,8 +19,8 @@ public:
 	Speed getSpeedY() const { return m_speedY; }
 
 	void setCenter(Point center) { m_center = center; }
-	void setX(int x) { m_center.x = x; }
-	void setY(int y) { m_center.y = y; }
+	void setX(float x) { m_center.x = x; }
+	void setY(float y) { m_center.y = y; }
 	void setRadius(Radius radius) { m_radius = radius; }
 	void setDirectionX(DirectionX dirX) { m_dirX = dirX; }
 	void setDirectionY(DirectionY dirY) { m_dirY = dirY; }
